refactor(front): rename some api slices
