+++ front/src/store/api/availabilityApiSlice.ts
@@ -26,7 +26,7 @@ export const extendedApiSlice = apiSlice.injectEndpoints({
         query: (id) => `profile/availabilities/${id}`,
         providesTags: (_result, _error, id) => [{ type: 'Availabilities', id }],
       }),
-      postAvailabilities: builder.mutation<Availability, CreateAvailabilityDto>(
+      postAvailability: builder.mutation<Availability, CreateAvailabilityDto>(
         {
           query: (body) => ({
             url: 'profile/availabilities',
@@ -36,7 +36,7 @@ export const extendedApiSlice = apiSlice.injectEndpoints({
           invalidatesTags: [{ type: 'Availabilities', id: 'LIST' }],
         },
       ),
-      patchAvailabilities: builder.mutation<
+      patchAvailability: builder.mutation<
         Availability,
         UpdateAvailabilityDto
       >({
@@ -50,7 +50,7 @@ export const extendedApiSlice = apiSlice.injectEndpoints({
           { type: 'Availabilities', id },
         ],
       }),
-      deleteAvailabilities: builder.mutation<void, string>({
+      deleteAvailability: builder.mutation<void, string>({
         query: (id) => ({
           url: `profile/availabilities/${id}`,
           method: 'DELETE',
@@ -67,7 +67,7 @@ export const extendedApiSlice = apiSlice.injectEndpoints({
 export const {
   useGetAvailabilitiesQuery,
   useGetAvailabilityQuery,
-  usePostAvailabilitiesMutation,
-  usePatchAvailabilitiesMutation,
-  useDeleteAvailabilitiesMutation,
+  usePostAvailabilityMutation,
+  usePatchAvailabilityMutation,
+  useDeleteAvailabilityMutation,
 } = extendedApiSlice;
